most code moved into abstract class
