+++ src/main/java/org/folio/service/tlr/TlrDataMigrationService.java
@@ -1,18 +1,9 @@
 package org.folio.service.tlr;
 
-import static io.vertx.core.Future.failedFuture;
 import static io.vertx.core.Future.succeededFuture;
-import static java.lang.String.format;
-import static java.lang.String.join;
 import static java.lang.System.currentTimeMillis;
-import static java.lang.System.lineSeparator;
-import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 import static java.util.stream.Collectors.toSet;
-import static java.util.stream.IntStream.range;
-import static org.apache.commons.lang3.time.DurationFormatUtils.formatDurationHMS;
-import static org.folio.rest.persist.PostgresClient.convertToPsqlStandard;
-import static org.folio.rest.tools.utils.TenantTool.tenantId;
 import static org.folio.support.JsonPropertyWriter.write;
 
 import java.util.ArrayList;
@@ -21,27 +12,16 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.StreamSupport;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.folio.okapi.common.ModuleId;
-import org.folio.okapi.common.SemVer;
-import org.folio.rest.client.OkapiClient;
 import org.folio.rest.jaxrs.model.TenantAttributes;
-import org.folio.rest.persist.Conn;
-import org.folio.rest.persist.PgUtil;
-import org.folio.rest.persist.PostgresClient;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import io.vertx.core.Context;
 import io.vertx.core.Future;
-import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import io.vertx.sqlclient.Row;
-import io.vertx.sqlclient.RowSet;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -54,10 +34,10 @@ public class TlrDataMigrationService extends AbstractRequestMigrationService{
   private static final String DEFAULT_UUID = "00000000-0000-4000-8000-000000000000";
 
   private static final String TLR_MIGRATION_MODULE_VERSION = "mod-circulation-storage-14.0.0";
-  private static final String REQUEST_TABLE = "request";
   private static final String ITEMS_STORAGE_URL = "/item-storage/items";
   private static final String HOLDINGS_STORAGE_URL = "/holdings-storage/holdings";
 
+  private static final String REQUEST_TABLE = "request";
   private static final String ITEM_REQUEST_LEVEL = "Item";
   private static final String REQUEST_LEVEL_KEY = "requestLevel";
   private static final String INSTANCE_ID_KEY = "instanceId";
@@ -70,40 +50,10 @@ public class TlrDataMigrationService extends AbstractRequestMigrationService{
 
   public TlrDataMigrationService(TenantAttributes attributes, Context context,
     Map<String, String> okapiHeaders) {
-      super(attributes, context, okapiHeaders);
+      super(attributes, context, okapiHeaders, REQUEST_TABLE, TLR_MIGRATION_MODULE_VERSION);
     }
 
-  public Future<Void> migrate() {
-    final long startTime = currentTimeMillis();
-
-    if (!shouldMigrate(TLR_MIGRATION_MODULE_VERSION)) {
-      return succeededFuture();
-    }
-
-    log.info("migration started, batch size " + BATCH_SIZE);
-
-    return getBatchCount()
-      .compose(this::migrateRequests)
-      .onSuccess(r -> log.info("Migration finished successfully"))
-      .onFailure(r -> log.error("Migration failed, rolling back the changes: {}", errorMessages))
-      .onComplete(r -> logDuration(startTime));
-  }
-
-  private Future<Void> migrateRequests(int batchCount) {
-    return postgresClient.withTrans(conn ->
-      chainFutures(buildBatches(batchCount, conn), this::processBatch)
-        .compose(r -> failIfErrorsOccurred())
-    );
-  }
-
-  private static Collection<Batch> buildBatches(int numberOfBatches, Conn connection) {
-    return range(0, numberOfBatches)
-      .boxed()
-      .map(num -> new Batch(num, connection))
-      .collect(toList());
-  }
-
-  private Future<Void> processBatch(Batch batch) {
+  public Future<Void> processBatch(Batch batch) {
     log.info("{} processing started", batch);
 
     return succeededFuture(batch)
@@ -117,35 +67,7 @@ public class TlrDataMigrationService extends AbstractRequestMigrationService{
       .recover(t -> handleError(batch, t));
   }
 
-  private Future<Batch> fetchRequests(Batch batch) {
-    return postgresClient.select(format("SELECT jsonb FROM %s.%s ORDER BY id LIMIT %d OFFSET %d",
-        schemaName, REQUEST_TABLE, BATCH_SIZE, batch.getBatchNumber() * BATCH_SIZE))
-      .onSuccess(r -> log.info("{} {} requests fetched", batch, r.size()))
-      .map(this::rowSetToRequestContexts)
-      .onSuccess(batch::setRequestMigrationContexts)
-      .map(batch);
-  }
-
-  private List<RequestMigrationContext> rowSetToRequestContexts(RowSet<Row> rowSet) {
-    return StreamSupport.stream(rowSet.spliterator(), false)
-      .map(row -> row.getJsonObject("jsonb"))
-      .map(RequestMigrationContext::from)
-      .collect(toList());
-  }
-
-  private Future<Batch> validateRequests(Batch batch) {
-    List<String> errors = batch.getRequestMigrationContexts()
-      .stream()
-      .map(this::validateRequest)
-      .flatMap(Collection::stream)
-      .collect(toList());
-
-    return errors.isEmpty()
-      ? succeededFuture(batch)
-      : failedFuture(join(lineSeparator(), errors));
-  }
-
-  private Collection<String> validateRequest(RequestMigrationContext context) {
+  public Collection<String> validateRequest(RequestMigrationContext context) {
     final JsonObject request = context.getOldRequest();
     final String requestId = context.getRequestId();
     final List<String> errors = new ArrayList<>();
@@ -209,12 +131,7 @@ public class TlrDataMigrationService extends AbstractRequestMigrationService{
       ctx.setInstanceId(holdingsRecordIdInstanceId.get(ctx.getHoldingsRecordId())));
   }
 
-  private void buildNewRequests(Batch batch) {
-    batch.getRequestMigrationContexts()
-      .forEach(this::buildNewRequest);
-  }
-
-  private void buildNewRequest(RequestMigrationContext context) {
+  public void buildNewRequest(RequestMigrationContext context) {
     String holdingsRecordId = context.getHoldingsRecordId();
     if (holdingsRecordId == null) {
       holdingsRecordId = DEFAULT_UUID;
@@ -253,34 +170,6 @@ public class TlrDataMigrationService extends AbstractRequestMigrationService{
     context.setNewRequest(migratedRequest);
   }
 
-  private Future<Void> updateRequests(Batch batch) {
-    if (!errorMessages.isEmpty()) {
-      log.info("{} batch update aborted - errors in previous batch(es) occurred", batch);
-      return succeededFuture();
-    }
-
-    List<JsonObject> migratedRequests = batch.getRequestMigrationContexts()
-      .stream()
-      .map(RequestMigrationContext::getNewRequest)
-      .collect(toList());
-
-    return batch.getConnection()
-      .updateBatch(REQUEST_TABLE, new JsonArray(migratedRequests))
-      .onSuccess(r -> log.info("{} all requests were successfully updated", batch))
-      .mapEmpty();
-  }
-
-  private Future<Void> failIfErrorsOccurred() {
-    return errorMessages.isEmpty()
-      ? succeededFuture()
-      : failedFuture(join(", ", errorMessages));
-  }
-
-  private static void logDuration(long startTime) {
-    String duration = formatDurationHMS(currentTimeMillis() - startTime);
-    log.info("Migration finished in {}", duration);
-  }
-
   private static boolean containsAll(JsonObject request, List<String> fieldNames) {
     return fieldNames.stream()
       .allMatch(request::containsKey);
@@ -289,27 +178,6 @@ public class TlrDataMigrationService extends AbstractRequestMigrationService{
   private static boolean containsAny(JsonObject request, List<String> fieldNames) {
     return fieldNames.stream()
       .anyMatch(request::containsKey);
-  }
-
-  private Future<Void> handleError(Batch batch, Throwable throwable) {
-    log.error("{} processing failed: {}", batch, throwable.getMessage());
-    errorMessages.add(throwable.getMessage());
-
-    return succeededFuture();
-  }
-
-  public static <T> Future<Void> chainFutures(Collection<T> list, Function<T, Future<Void>> method) {
-    return list.stream().reduce(succeededFuture(),
-      (acc, item) -> acc.compose(v -> method.apply(item)),
-      (a, b) -> succeededFuture());
-  }
-
-  private static SemVer moduleVersionToSemVer(String version) {
-    try {
-      return new SemVer(version);
-    } catch (IllegalArgumentException ex) {
-      return new ModuleId(version).getSemVer();
-    }
   }
 
   @Getter
